changing disk_used_percentage setting
Originally the variable is set by looking for the data directory's size then grepping it against disk usage size. However, if there is a difference like 1.5T vs 1.6T, then the 1.5T can't be found in the output of df -lh and the rest of the piping breaks. 

In the proposed code change, we are instead looking for the disk name where the datadir is, and from there we look for disk name in the output of df -lh. This should be more robust as disk names don't change unlike size and it's pretty much a sure thing to find the disk name that contains the datadir.